# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open API documentation for project Album-singers",
        "contact": {
            "email": "galina@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "songsAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Create yet another song record",
                "operationId": "songCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Song name",
                                        "type": "string",
                                        "example": "Example"
                                    },
                                    "singer_id": {
                                        "description": "Singer ID",
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Song name",
                                            "type": "string",
                                            "example": "Name of Song"
                                        },
                                        "singer": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SingerShowResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "songsOne",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of song",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "name": {
                                            "description": "Name of key for storring",
                                            "type": "string",
                                            "example": "Name of Song"
                                        },
                                        "singer": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SingerShowResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Delete song by ID",
                "operationId": "songsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of song",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successful"
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Songs"
                ],
                "summary": "Update yet another song record",
                "operationId": "songUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of song",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Song name",
                                        "type": "string",
                                        "example": "Name of song"
                                    },
                                    "singer_id": {
                                        "description": "Singer ID",
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Song name",
                                            "type": "string",
                                            "example": "Name of Song"
                                        },
                                        "singer": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SingerShowResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SingerShowResource": {
                "title": "Singer model",
                "description": "Some simple request createa as example",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "5"
                    },
                    "name": {
                        "description": "The name",
                        "type": "string",
                        "example": "Alex"
                    }
                },
                "type": "object"
            }
        }
    },
...
}